Add files via upload
The html file contains the basic structure of the webpage which includes the boxes to mark 'x' or 'o'. And also includes a reset button.
Then the CSS file includes the styling code for the page and finally the javascript page which has the logic for reset and game winning conditions check which is alerted in the game.